# Remove one level of indirection