make it CPU cache friendly