@eranchetz 2016-08-14
@redbaron Can you please add benchmarks for this change?