@redbaron 2016-08-14
there is an empirical evidence - why waste 32 cache lines for each shard struct, when they all can be compact together in `32*sizeof(shard)/64` lines.

I'll add benchmarks soon, but just to show that there is no performance degradation. I expect improvement not to be seen  in synthetic runs as there is not cache pressure when single library is tested. In real apps & virtualised environments CPU  cache is a very scarce resource, that is where I expect this change to have some effect.