@eranchetz 2016-08-14
Great, I am expecting your benchmarks.
Thanks a lot for your support and input to this project!

On Sun, Aug 14, 2016 at 4:07 PM, Maxim Ivanov <notifications@github.com>
wrote:

> there is an empirical evidence - why waste 32 cache lines for each shard
> struct, when they all can be compact together in 32*sizeof(shard)/64
> lines.
>
> I'll add benchmarks soon, but just to show that there is no performance
> degradation. I expect improvement not to be seen in synthetic runs as there
> is not cache pressure when single library is tested. In real apps &
> virtualised environments CPU cache is a very scarce resource, that is where
> I expect this change to have some effect.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/streamrail/concurrent-map/pull/34#issuecomment-239672617>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAVy5gRnDNmi73c8mYHVOJQgxs7G37RVks5qfxMSgaJpZM4Jj2nE>
> .
>