@redbaron 2016-08-14
Before:
```
BenchmarkItems-4                         	     200	   8427049 ns/op
BenchmarkMarshalJson-4                   	      50	  26016675 ns/op
BenchmarkStrconv-4                       	20000000	       130 ns/op
BenchmarkSingleInsertAbsent-4            	 1000000	      1677 ns/op
BenchmarkSingleInsertPresent-4           	 5000000	       262 ns/op
BenchmarkMultiInsertDifferent_1_Shard-4  	  500000	      4839 ns/op
BenchmarkMultiInsertDifferent_16_Shard-4 	  500000	      4665 ns/op
BenchmarkMultiInsertDifferent_32_Shard-4 	  500000	      4648 ns/op
BenchmarkMultiInsertDifferent_256_Shard-4	  300000	      5553 ns/op
BenchmarkMultiInsertSame-4               	  500000	      2680 ns/op
BenchmarkMultiGetSame-4                  	 1000000	      1020 ns/op
BenchmarkMultiGetSetDifferent_1_Shard-4  	  300000	      5972 ns/op
BenchmarkMultiGetSetDifferent_16_Shard-4 	  300000	      5723 ns/op
BenchmarkMultiGetSetDifferent_32_Shard-4 	  300000	      5629 ns/op
BenchmarkMultiGetSetDifferent_256_Shard-4	  300000	      5596 ns/op
BenchmarkMultiGetSetBlock_1_Shard-4      	  300000	      4342 ns/op
BenchmarkMultiGetSetBlock_16_Shard-4     	  300000	      4511 ns/op
BenchmarkMultiGetSetBlock_32_Shard-4     	  300000	      4291 ns/op
BenchmarkMultiGetSetBlock_256_Shard-4    	  500000	      4078 ns/op
BenchmarkKeys-4                          	     500	   3358915 ns/op
ok  	_/home/rossmohax/repos/external/concurrent-map	37.241s
```

After:
```
PASS
BenchmarkItems-4                         	     200	   8275102 ns/op
BenchmarkMarshalJson-4                   	      50	  25469559 ns/op
BenchmarkStrconv-4                       	20000000	       128 ns/op
BenchmarkSingleInsertAbsent-4            	 1000000	      1765 ns/op
BenchmarkSingleInsertPresent-4           	 5000000	       256 ns/op
BenchmarkMultiInsertDifferent_1_Shard-4  	  500000	      4907 ns/op
BenchmarkMultiInsertDifferent_16_Shard-4 	  500000	      4602 ns/op
BenchmarkMultiInsertDifferent_32_Shard-4 	  500000	      4609 ns/op
BenchmarkMultiInsertDifferent_256_Shard-4	  300000	      5601 ns/op
BenchmarkMultiInsertSame-4               	  500000	      2726 ns/op
BenchmarkMultiGetSame-4                  	 1000000	      1005 ns/op
BenchmarkMultiGetSetDifferent_1_Shard-4  	  300000	      5893 ns/op
BenchmarkMultiGetSetDifferent_16_Shard-4 	  300000	      5754 ns/op
BenchmarkMultiGetSetDifferent_32_Shard-4 	  300000	      5655 ns/op
BenchmarkMultiGetSetDifferent_256_Shard-4	  300000	      5625 ns/op
BenchmarkMultiGetSetBlock_1_Shard-4      	  300000	      4341 ns/op
BenchmarkMultiGetSetBlock_16_Shard-4     	  300000	      4583 ns/op
BenchmarkMultiGetSetBlock_32_Shard-4     	  300000	      4284 ns/op
BenchmarkMultiGetSetBlock_256_Shard-4    	  500000	      4120 ns/op
BenchmarkKeys-4                          	     500	   3282857 ns/op
ok  	_/home/rossmohax/repos/external/concurrent-map	37.215s
```